requirements.txt
discord.py
requests
pytube
